Updates utilities
