+++ src/models/Popup.js
@@ -9,9 +9,8 @@ import {
   statusAllow
 } from "../constants"
 import {
-  getContrast,
+  addCustomStylesheet,
   getCookie,
-  getHoverColor,
   hash,
   interpolateString,
   isMobile,
@@ -20,7 +19,7 @@ import {
   setCookie,
   throttle,
   traverseDOMPath,
-} from "../utils.js"
+} from "../utils"
 
 export default class Popup extends Base {
   constructor( options ) {
@@ -495,136 +494,11 @@ export default class Popup extends Base {
     this.customStyleSelector = isValid ? selector : null
 
     if (isValid) {
-      this.addCustomStyle(hashStr, palette, '.' + selector)
+      addCustomStylesheet(hashStr, palette, '.' + selector)
     }
     return isValid
   }
 
-  addCustomStyle(hashStr, palette, prefix) {
-    // only add this if a style like it doesn't exist
-    if (this.customStyles[hashStr]) {
-      // custom style already exists, so increment the reference count
-      ++this.customStyles[hashStr].references
-      return
-    }
-
-    const colorStyles = {}
-    const { popup, button, highlight, saveButton } = palette
-
-    // needs background colour, text and link will be set to black/white if not specified
-    if (popup) {
-      // assumes popup.background is set
-      popup.text = popup.text
-        ? popup.text
-        : getContrast(popup.background)
-      popup.link = popup.link ? popup.link : popup.text
-      colorStyles[prefix + ' .cc-tooltip, ' + prefix + ' .cc-tooltip:after'] = [
-        'color: ' + popup.text,
-        'background-color: ' + popup.background
-      ]
-      colorStyles[prefix + '.cc-window'] = [
-        'color: ' + popup.text,
-        'background-color: ' + popup.background
-      ]
-      colorStyles[prefix + '.cc-revoke'] = [
-        'color: ' + popup.text,
-        'background-color: ' + popup.background
-      ]
-      colorStyles[
-        prefix +
-          ' .cc-link,' +
-          prefix +
-          ' .cc-link:active,' +
-          prefix +
-          ' .cc-link:visited'
-      ] = ['color: ' + popup.link]
-
-      if (button) {
-        // assumes button.background is set
-        button.text = button.text
-          ? button.text
-          : getContrast(button.background)
-        button.border = button.border ? button.border : 'transparent'
-        colorStyles[prefix + ' .cc-btn'] = [
-          'color: ' + button.text,
-          'border-color: ' + button.border,
-          'background-color: ' + button.background
-        ]
-
-        if (button.padding) {
-          colorStyles[prefix + ' .cc-btn'].push('padding: ' + button.padding)
-        }
-
-        if (button.background != 'transparent') {
-          colorStyles[
-            prefix + ' .cc-btn:hover, ' + prefix + ' .cc-btn:focus'
-          ] = [
-            'background-color: ' +
-              (button.hover || getHoverColor(button.background))
-          ]
-        }
-
-        if (highlight) {
-          //assumes highlight.background is set
-          highlight.text = highlight.text
-            ? highlight.text
-            : getContrast(highlight.background)
-          highlight.border = highlight.border
-            ? highlight.border
-            : 'transparent'
-          colorStyles[prefix + ' .cc-highlight .cc-btn:first-child'] = [
-            'color: ' + highlight.text,
-            'border-color: ' + highlight.border,
-            'background-color: ' + highlight.background
-          ]
-        } else {
-          // sets highlight text color to popup text. background and border are transparent by default.
-          colorStyles[prefix + ' .cc-highlight .cc-btn:first-child'] = [
-            'color: ' + popup.text
-          ]
-        }
-      }
-
-      if ( saveButton ) {
-        colorStyles[prefix + ' .cc-btn.cc-save'] = [
-          'color: ' + saveButton.text,
-          'border-color: ' + saveButton.border,
-          'background-color: ' + saveButton.background
-        ]
-      }
-    }
-    // this will be interpretted as CSS. the key is the selector, and each array element is a rule
-    const style = document.createElement('style')
-    document.head.appendChild(style)
-
-    // custom style doesn't exist, so we create it
-    this.customStyles[hashStr] = {
-      references: 1,
-      element: style.sheet
-    }
-
-    let ruleIndex = -1
-    for (let prop in colorStyles) {
-      if (colorStyles.hasOwnProperty(prop)) {
-        style.sheet.insertRule(
-          prop + '{' + colorStyles[prop].join(';') + '}',
-          ++ruleIndex
-        )
-      }
-    }
-  }
-
-  removeCustomStyle(palette) {
-    const hashStr = hash(JSON.stringify(palette))
-    const customStyle = this.customStyles[hashStr]
-    if (customStyle && !--customStyle.references) {
-      const styleNode = customStyle.element.ownerNode
-      if (styleNode && styleNode.parentNode) {
-        styleNode.parentNode.removeChild(styleNode)
-      }
-      this.customStyles[hashStr] = null
-    }
-  }
 
   getEventPath( event ) {
     const path = event.composedPath ? event.composedPath() : (function ( arr, element ) {
